Burst wave testing in progress
Not working yet, considering a re-write of functionality. The way the _get_chan_n_waveform function is written right now is forcing the code to have a lot of opportunities to just return the inactive waveform. It might make more sense to just make two getters, one for each waveform type instead of getting whichever is active.