Basic and burst mode for function generators (#32)
* Added getters/setters for burst mode and modified get/set waveform functions to handle both basic and burst mode

* Burst wave testing in progress

Not working yet, considering a re-write of functionality. The way the _get_chan_n_waveform function is written right now is forcing the code to have a lot of opportunities to just return the inactive waveform. It might make more sense to just make two getters, one for each waveform type instead of getting whichever is active.

* Created separate get/set waveform functions for basic and burst modes

* Update generators.py

Final updates for burst mode

---------

Co-authored-by: Celena Michaud <47991498+celenamichaud@users.noreply.github.com>